pyproject.toml
[tool.poetry]
name = "voice-to-voice-ai"
version = "0.1.0"
description = ""
authors = ["frzyc <frzyc@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
rich = "^13.7.1"
openai = "^1.34.0"
azure-cognitiveservices-speech = "^1.38.0"
keyboard = "^0.13.5"
obs-websocket-py = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
